fix: move Stylesheet type to `@lwc/shared`
