+++ packages/@lwc/engine-core/src/framework/check-version-mismatch.ts
@@ -9,9 +9,9 @@ import { isNull, LWC_VERSION, LWC_VERSION_COMMENT_REGEX } from '@lwc/shared';
 import { logError } from '../shared/logger';
 
 import { Template } from './template';
-import { Stylesheet } from './stylesheet';
 import { LightningElementConstructor } from './base-lightning-element';
 import { report, ReportingEventId } from './reporting';
+import type { Stylesheet } from '@lwc/shared';
 
 let warned = false;
