+++ packages/@lwc/engine-core/src/framework/hot-swaps.ts
@@ -12,9 +12,10 @@ import { LightningElementConstructor } from './base-lightning-element';
 import { Template } from './template';
 import { markComponentAsDirty } from './component';
 import { isTemplateRegistered } from './secure-template';
-import { Stylesheet, Stylesheets, unrenderStylesheet } from './stylesheet';
+import { unrenderStylesheet } from './stylesheet';
 import { assertNotProd, flattenStylesheets } from './utils';
 import { WeakMultiMap } from './weak-multimap';
+import type { Stylesheet, Stylesheets } from '@lwc/shared';
 
 let swappedTemplateMap: WeakMap<Template, Template> = /*@__PURE__@*/ new WeakMap();
 let swappedComponentMap: WeakMap<LightningElementConstructor, LightningElementConstructor> =
