+++ packages/@lwc/ssr-runtime/package.json
@@ -44,7 +44,6 @@
         }
     },
     "devDependencies": {
-        "@lwc/engine-core": "8.3.0",
         "@lwc/shared": "8.3.0"
     }
 }
